Add flags. Temperary move nwidger to kaicheng for dev purpose.
