+++ main.go
@@ -8,8 +8,8 @@ import (
 	"flag"
 
 	"github.com/mitchellh/go-homedir"
-	"github.com/nwidger/nintengo/http"
-	"github.com/nwidger/nintengo/nes"
+	"github.com/kaicheng/nintengo/http"
+	"github.com/kaicheng/nintengo/nes"
 	"gopkg.in/yaml.v2"
 )
 
@@ -37,6 +37,8 @@ func main() {
 	flag.StringVar(&options.CPUProfile, "cpu-profile", "", "write CPU profile to file")
 	flag.StringVar(&options.MemProfile, "mem-profile", "", "write memory profile to file")
 	flag.StringVar(&options.HTTPAddress, "http", "", "HTTP service address (e.g., ':6060')")
+	flag.StringVar(&options.Listen, "l", "", "Listen at address as master (e.g., ':8080')")
+	flag.StringVar(&options.Connect, "c", "", "Connect to address as slave (e.g., 'localhost:8080')")
 	flag.Parse()
 
 	filename, err := homedir.Expand("~/.nintengorc")
