+++ nes/cnrom.go
@@ -3,7 +3,7 @@ package nes
 import (
 	"fmt"
 
-	"github.com/nwidger/nintengo/rp2ago3"
+	"github.com/kaicheng/nintengo/rp2ago3"
 )
 
 type CNROMRegisters struct {
