+++ nes/nes.go
@@ -14,9 +14,9 @@ import (
 
 	"archive/zip"
 
-	"github.com/nwidger/nintengo/m65go2"
-	"github.com/nwidger/nintengo/rp2ago3"
-	"github.com/nwidger/nintengo/rp2cgo2"
+	"github.com/kaicheng/nintengo/m65go2"
+	"github.com/kaicheng/nintengo/rp2ago3"
+	"github.com/kaicheng/nintengo/rp2cgo2"
 )
 
 //go:generate stringer -type=StepState
