+++ nes/nes_test.go
@@ -3,7 +3,7 @@ package nes
 import (
 	"testing"
 
-	"github.com/nwidger/nintengo/rp2cgo2"
+	"github.com/kaicheng/nintengo/rp2cgo2"
 )
 
 func TestVerticalMirroring(t *testing.T) {
