+++ nes/rom.go
@@ -11,8 +11,8 @@ import (
 
 	"strings"
 
-	"github.com/nwidger/nintengo/rp2ago3"
-	"github.com/nwidger/nintengo/rp2cgo2"
+	"github.com/kaicheng/nintengo/rp2ago3"
+	"github.com/kaicheng/nintengo/rp2cgo2"
 )
 
 //go:generate stringer -type=Region
