+++ rp2ago3/dma.go
@@ -1,6 +1,6 @@
 package rp2ago3
 
-import "github.com/nwidger/nintengo/m65go2"
+import "github.com/kaicheng/nintengo/m65go2"
 
 const NO_PENDING uint32 = 0xffffffff
